+++ Tuist/ProjectDescriptionHelpers/Module.swift
@@ -73,6 +73,7 @@ public enum Module: String, CaseIterable {
     case inspectCommand = "TuistInspectCommand"
     case android = "TuistAndroid"
     case machineMetrics = "TuistMachineMetrics"
+    case appleArchiver = "TuistAppleArchiver"
 
     func forceStaticLinking() -> Bool {
         return Environment.forceStaticLinking.getBoolean(default: false)
@@ -465,7 +466,7 @@ public enum Module: String, CaseIterable {
             moduleTags.append("domain:plugins")
         case .simulator, .xcActivityLog, .git, .rootDirectoryLocator,
             .process, .ci, .cas, .casAnalytics, .launchctl, .xcResultService, .xcodeProjectOrWorkspacePathLocator,
-            .http, .har, .configLoader, .machineMetrics:
+            .http, .har, .configLoader, .machineMetrics, .appleArchiver:
             moduleTags.append("domain:infrastructure")
         case .cacheCommand, .authCommand, .envKey, .versionCommand,
              .accountCommand, .organizationCommand, .projectCommand, .bundleCommand,
@@ -719,6 +720,8 @@ public enum Module: String, CaseIterable {
                     .target(name: Module.xcodeProjectOrWorkspacePathLocator.targetName),
                     .target(name: Module.xcResultService.targetName),
                     .target(name: Module.cas.targetName),
+                    .target(name: Module.appleArchiver.targetName, condition: .when([.macos])),
+                    .target(name: Module.casAnalytics.targetName),
                     .target(name: Module.launchctl.targetName),
                     .target(name: Module.machineMetrics.targetName),
                     .target(name: Module.oidc.targetName),
@@ -1031,9 +1034,8 @@ public enum Module: String, CaseIterable {
                 ]
             case .casAnalytics:
                 [
-                    .target(name: Module.support.targetName),
                     .target(name: Module.environment.targetName),
-                    .external(name: "FileSystem"),
+                    .external(name: "CASAnalyticsDatabase"),
                 ]
             case .launchctl:
                 [
@@ -1337,6 +1339,11 @@ public enum Module: String, CaseIterable {
                     .external(name: "FileSystem"),
                     .external(name: "TSCBasic"),
                 ]
+            case .appleArchiver:
+                [
+                    .external(name: "Path"),
+                    .external(name: "Mockable"),
+                ]
             case .shareCommand:
                 [
                     .target(name: Module.server.targetName),
@@ -1408,6 +1415,7 @@ public enum Module: String, CaseIterable {
                     .target(name: Module.xcodeProjectOrWorkspacePathLocator.targetName),
                     .target(name: Module.xcResultService.targetName),
                     .target(name: Module.machineMetrics.targetName),
+                    .target(name: Module.casAnalytics.targetName),
                     .target(name: Module.ci.targetName, condition: .when([.macos])),
                     .target(name: Module.process.targetName, condition: .when([.macos])),
                     .target(name: Module.config.targetName),
@@ -1444,6 +1452,11 @@ public enum Module: String, CaseIterable {
                     .target(name: Module.testing.targetName),
                     .target(name: Module.environment.targetName),
                     .target(name: Module.environmentTesting.targetName),
+                    .external(name: "FileSystem"),
+                    .external(name: "FileSystemTesting"),
+                ]
+            case .appleArchiver:
+                [
                     .external(name: "FileSystem"),
                     .external(name: "FileSystemTesting"),
                 ]
@@ -1583,6 +1596,7 @@ public enum Module: String, CaseIterable {
                     .target(name: Module.xcodeProjectOrWorkspacePathLocator.targetName),
                     .target(name: Module.launchctl.targetName),
                     .target(name: Module.machineMetrics.targetName),
+                    .target(name: Module.casAnalytics.targetName),
                     .target(name: Module.oidc.targetName),
                     .target(name: Module.http.targetName),
                     .target(name: Module.cas.targetName),